+++ src/frontend/src/pages/provider/bookings.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect, useMemo } from "react";
-import { useNavigate} from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 import BottomNavigation from "../../components/provider/BottomNavigationNextjs";
 import ProviderBookingItemCard from "../../components/provider/ProviderBookingItemCard";
 import {